refactored format vehicle additional defects to defect service
